requirements.txt
amqp==5.1.1
asgiref==3.5.1
async-timeout==4.0.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.7
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.29.0
cryptography==37.0.2
cssutils==2.4.2
Deprecated==1.2.13
Django==4.0.4
django-celery-beat==2.3.0
django-compression-middleware==0.4.1
django-crispy-forms==1.14.0
django-inlinecss==0.3.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-phonenumber-field==6.3.0
django-postmark==0.1.6
django-redis==5.2.0
django-sslserver==0.22
django-timezone-field==5.0
django-utils-six==2.0
future==0.18.2
gunicorn==20.1.0
heroku==0.1.4
httplib2==0.20.4
idna==3.3
kombu==5.2.4
packaging==21.3
Pillow==9.1.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pycparser==2.21
PyMySQL==1.0.2
pynliner==0.8.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==1.5
python-dotenv==0.20.0
pytz==2022.1
rabbitmq==0.2.0
redis==4.3.4
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.1.0
wrapt==1.14.1
zstandard==0.18.0
